datefixR: Standardize Dates in Different Formats or with Missing Data

There are many different formats dates are commonly represented with: the order of day, month, or year can differ, different separators ("-", "/", or whitespace) can be used, months can be numerical, names, or abbreviations and year given as two digits or four. datefixR takes dates in all these different formats and converts them to R's built-in date class. If datefixR cannot standardize a date, such as because it is too malformed, then the user is told which date cannot be standardized and the corresponding ID for the row. datefixR also allows the imputation of missing days and months with user-controlled behavior.

Get started by reading vignette("datefixR")

See also

Author

Maintainer: Nathan Constantine-Cooke nathan.constantine-cooke@ed.ac.uk (ORCID)

Other contributors:

Jonathan Kitt jonathan.kitt@protonmail.com [contributor, translator]

Antonio J. Pérez-Luque ajpelu@gmail.com (ORCID) [contributor, translator]

Daniel Possenriede possenriede+r@gmail.com (ORCID) [contributor, translator]

Michal Lauer michal.lauer.25@gmail.com [contributor, translator]

Kaique dos S. Alves kaiquedsalves@gmail.com (ORCID) [reviewer]

Al-Ahmadgaid B. Asaad alahmadgaid@gmail.com (ORCID) [reviewer]

Anatoly Tsyplenkov atsyplenkov@gmail.com (ORCID) [contributor, translator]

Chitra M. Saraswati chitra.m.saraswati@gmail.com (ORCID) [contributor, translator]